crystal: add interpreter to build
This commit adds the REPL interpreter to the commit by setting
`interpreter=1`. This will allow usage of `icr` through the command
`crystal interactive` or `crystal i`.

Limitation is that `pcre2` doesn't link correctly when in interpreter
mode using pkg-config? I am not sure what is triggering this, but the
solution I came to was to manually point this library to
`${pcre2.out}/lib`.